endpoint: refactor Docker + IPVLAN interactions
The endpoint's internals were leaked in the `finishIpvlanInit`
function. Refactor this function to be part of the endpoint, and hide the
implementation of the actual IPVLAN + netNS interactions behind an interface as
to not leak implementation of the datapath itself into `pkg/endpoint`.

Signed-off by: Ian Vernon <ian@cilium.io>